requirements.txt
streamlit
opencv-python
scikit-image
joblib
numpy
pillow
